Got rid of bloated main file
