+++ sensor_c/inc/main.h
@@ -1,0 +1,10 @@
+#ifndef MAIN_H_
+#define MAIN_H_
+
+#include "sys_control.h"
+#include "analog.h"
+#include "serial.h"
+#include "clock.h"
+
+
+#endif /* MAIN_H_ */
